refactor(utils): reorganize utils
