+++ src/index.ts
@@ -3,12 +3,5 @@ export { Pagination, PaginationDefaults, PaginationKwargs, IPagination } from ".
 export type { PaginationFilters } from "./api"
 export { createApi, createCustomServiceCall, createPaginatedServiceCall } from "./api"
 export { createCollectionManager } from "./collection-manager"
-export type {
-  GetInferredFromRaw,
-  FromApiCall,
-  ToApiCall,
-  GetInferredRecursiveShape,
-  GetRecursiveZodShape,
-} from "./utils"
-export { createApiUtils, recursiveShapeToValidZodRawShape } from "./utils"
-export { parseResponse } from "./response"
+export type { GetInferredFromRaw, FromApiCall, ToApiCall } from "./utils"
+export { createApiUtils, parseResponse } from "./utils"
